Fix bug with selection when getting features by ID
- Introduce new base event type "SelectiveEvent" which is the new super class of QueryEvent & GetFeaturesByIdEvent to not have a redundant definition of "selection"
- Adjusted "GetFeatures" query-runner to support selection for all "SelectiveEvents"
- Also added according integration test
